requirements.txt
python-telegram-bot==20.7
requests==2.31.0
speedtest-cli==2.1.3
google-generativeai==0.3.2

